+++ objc2/src/macros.rs
@@ -1,3 +1,9 @@
+mod declare_class;
+mod extern_class;
+#[doc(hidden)]
+#[cfg(feature = "foundation")]
+pub mod ns_string;
+
 /// Gets a reference to a [`Class`] from the given name.
 ///
 /// [`Class`]: crate::runtime::Class